Moving Topology and Notification to the interface
